* develop: (63 commits)
  auto deploy from ci pipeline bucket because auto deploy is now done before promotion
  consolidate settings retrieval via GetSettings
  handle NotSupported/NotImplemented/VMNotFound errors in external cpis
  simplified external_cpi_spec.rb
  properly reset external cpi usage in integration tests
  fix typo
  suffix blobstores with blobstore
  renamed sha1Verifiable to sha1VerifiableBlobstore
  retry getting blob in go agent
  Fix flaky network configuration integration test
  moved platform specific network-change-preparation code into platform
  do not include kernel buffers/cached memory amount in mem used vital value
  formatting
  added integration spec to make sure dynamic network settings are properly passed in to job templates
  re-render job templates after resolving dynamic ips
  make job renderer available from VmStateApplier
  remove extra indirection
  introduce vm_updater to clean up instance_updater
  added integration test for changing vm networking with cpi that does not support certain changes without recreating vm
  formatting
  ...

+++ bosh-stemcell/spec/support/os_image.rb
@@ -10,6 +10,6 @@ RSpec.configure do |config|
   end
 
   config.after(:all, example_group: { file_path: /(spec|support)\/os_image/ }) do
-    FileUtils.rm_rf(@os_image_dir)
+    FileUtils.rm_rf(@os_image_dir) if ENV['OS_IMAGE']
   end
 end
